Moved Answer: Note that there is a problem in performCorrection for question 8.
