+++ profiles_api/urls.py
@@ -7,7 +7,7 @@ from profiles_api import views
 from profiles_api.topic import topic_api_view
 from profiles_api.subtopic import subtopic_api_view
 from profiles_api.question import question_api_view
-
+from profiles_api.answer import answer_api_view
 
 
 router = DefaultRouter()
@@ -17,7 +17,7 @@ router.register('feed', views.UserProfileFeedViewSet)
 router.register('question', question_api_view.QuestionViewSet)
 router.register('subtopic', subtopic_api_view.SubtopicViewSet)
 router.register('topic', topic_api_view.TopicViewSet)
-router.register('answer', views.AnswerViewSet)
+router.register('answer', answer_api_view.AnswerViewSet)
 router.register('test', views.TestViewSet)
 router.register('completedTest', views.CompletedTestViewSet)
 router.register('theoryPage', views.TheoryPageViewSet)
@@ -29,6 +29,6 @@ urlpatterns = [
     path('custom-subtopic/', subtopic_api_view.CustomSubtopicView.as_view()),
     path('custom-question/', question_api_view.QuestionView.as_view()),
     path('custom-topic/', topic_api_view.TopicView.as_view()),
-    path('custom-answer/', views.AnswerView.as_view()),
+    path('custom-answer/', answer_api_view.AnswerView.as_view()),
     path('', include(router.urls))
 ]
